requirements.txt
streamlit==1.30.0
langchain==0.0.173
openai==1.0.0
python-dotenv==1.0.0
pydantic==1.10.7